Update Birthday_congratulation.gs
Now Install creates the labels, the filter and the Trigger
Searching for new Threads to congratulate wont be done with the search method in GmailApp, but with the getUserLabelByName method in GmailApp.
Is it better? Maybe. It isn't as complex. The search method requires you to set the labels with -, as oposed by the convenction in the other methods or inside of gmail